+++ src/imio/dashboard/tests/test_columns.py
@@ -1,7 +1,14 @@
 # -*- coding: utf-8 -*-
 
+from zope.component import getUtility
+from zope.intid.interfaces import IIntIds
+from z3c.relationfield.relation import RelationValue
+
+from plone import api
+
 from imio.dashboard.columns import ActionsColumn
 from imio.dashboard.columns import PrettyLinkColumn
+from imio.dashboard.columns import RelationPrettyLinkColumn
 from imio.dashboard.testing import IntegrationTestCase
 
 
@@ -33,3 +40,33 @@ class TestColumns(IntegrationTestCase):
         self.assertIn("/edit", rendered_column)
         self.assertIn("javascript:confirmDeleteObject", rendered_column)
         self.assertIn("history.gif", rendered_column)
+
+    def test_RelationPrettyLinkColumn(self):
+        """Test the RelationPrettyLinkColumn, it will render IPrettyLink.getLink."""
+        table = self.faceted_table
+        column = RelationPrettyLinkColumn(self.portal, self.portal.REQUEST, table)
+        fold1 = api.content.create(container=self.portal, type='Folder', id='fold1', title="Folder 1")
+        fold2 = api.content.create(container=self.portal, type='Folder', id='fold2', title="Folder 2")
+        intids = getUtility(IIntIds)
+        rel1 = RelationValue(intids.getId(fold1))
+        rel2 = RelationValue(intids.getId(fold2))
+        tt = api.content.create(container=self.portal, type='testingtype', id='testingtype',
+                                title='My testing type', rel_item=rel1, rel_items=[rel1, rel2])
+        brain = self.portal.portal_catalog(UID=tt.UID())[0]
+        column.attrName = 'rel_item'
+        self.assertEqual(u"<a class='pretty_link' title='' href='http://nohost/plone/fold1' target='_self'>"
+                         "<span class='pretty_link_content'>Folder 1</span></a>",
+                         column.renderCell(brain))
+        column.params = {'showContentIcon': True}
+        self.assertEqual(u"<a class='pretty_link contenttype-Folder' title='' href='http://nohost/plone/fold1' "
+                         "target='_self'><span class='pretty_link_content'>Folder 1</span></a>",
+                         column.renderCell(brain))
+        column.params = {}
+        column.attrName = 'rel_items'
+        self.assertEqual(u"<ul>\n<li><a class='pretty_link' title='' href='http://nohost/plone/fold1' target='_self'>"
+                         "<span class='pretty_link_content'>Folder 1</span></a></li>\n"
+                         "<li><a class='pretty_link' title='' href='http://nohost/plone/fold2' target='_self'>"
+                         "<span class='pretty_link_content'>Folder 2</span></a></li>\n</ul>",
+                         column.renderCell(brain))
+        # a pretty_link class is defined for the td
+        self.assertEquals(column.cssClasses, {'td': 'pretty_link', 'th': 'th_header_rel_items'})